Partial setup of lua playdate game

+++ build.fish
@@ -1,9 +1,6 @@
 #!/usr/bin/env fish
 
-# Set ARM toolchain paths for Linux
-set -x ARM_NONE_EABI_GCC_PATH /usr/lib/gcc/arm-none-eabi/15
-set -x ARM_NONE_EABI_SYSROOT_PATH /usr/arm-none-eabi
-
+set PLAYDATE_SDK_PATH "/root/PlaydateSDK/PlaydateSDK-3.0.2/"
 # Set product name
 set PRODUCT_NAME Gutwound
 
@@ -13,16 +10,31 @@ end
 
 trap cleanup EXIT
 
-# Build the package, filtering out pointless library version warnings
-swift package pdc 2>&1 | grep -v "libtinfo.so.6: no version information" | grep -v "libxml2.so.2: no version information" | grep -v "libcurl.so.4: no version information"
+# Check if PLAYDATE_SDK_PATH is set
+if not set -q PLAYDATE_SDK_PATH
+    echo "Error: PLAYDATE_SDK_PATH environment variable is not set"
+    echo "Please set it to your Playdate SDK installation directory"
+    exit 1
+end
+
+# Build the package using pdc (Playdate Compiler)
+echo "Building $PRODUCT_NAME.pdx..."
+$PLAYDATE_SDK_PATH/bin/pdc source .builds/$PRODUCT_NAME.pdx
+
+if test $status -ne 0
+    echo "Build failed!"
+    exit 1
+end
+
+echo "Build successful!"
 
 # Create a symbolic link to the compiled pdx in the Playdate Simulator Games dir
 # This allows browsing the pdx as a sideloaded game, helpful for checking icons
 if test -L $PLAYDATE_SDK_PATH/Disk/Games/$PRODUCT_NAME.pdx
     rm -rf $PLAYDATE_SDK_PATH/Disk/Games/$PRODUCT_NAME.pdx
 end
-ln -s (pwd)/.build/plugins/PDCPlugin/outputs/$PRODUCT_NAME.pdx $PLAYDATE_SDK_PATH/Disk/Games/$PRODUCT_NAME.pdx
+ln -s (pwd)/.builds/$PRODUCT_NAME.pdx $PLAYDATE_SDK_PATH/Disk/Games/$PRODUCT_NAME.pdx
 
-cd .build
 # Launch the simulator (note: may need webkit2gtk-4.1 library installed on Linux)
-$PLAYDATE_SDK_PATH/bin/PlaydateSimulator $PLAYDATE_SDK_PATH/Disk/Games/$PRODUCT_NAME.pdx
+echo "Launching Playdate Simulator..."
+$PLAYDATE_SDK_PATH/bin/PlaydateSimulator .builds/$PRODUCT_NAME.pdx
